requirements.txt
tabulate
stanfordnlp
spacy-transformers
